srcs/requirements/tournament_service/app/requirements.txt
Django==5.0.1
psycopg2-binary==2.9.6
djangorestframework==3.12.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==3.5.0
pyjwt==2.8.0
django-environ==0.11.2
